+++ package.json
@@ -51,6 +51,7 @@
     "markdown-it": "^13.0.1",
     "metro-minify-terser": "^0.74.0",
     "patch-package": "^6.5.1",
+    "query-string": "^8.1.0",
     "randombytes": "^2.1.0",
     "react": "18.2.0",
     "react-native": "0.70.6",
