+++ src/event-actions/on-app-started.ts
@@ -1,15 +1,33 @@
-import {Linking} from 'react-native';
+import NetInfo, {NetInfoState} from '@react-native-community/netinfo';
+import {AppState, Linking} from 'react-native';
 
 import {app} from '@app/contexts';
 import {Events} from '@app/events';
+import {hideModal, showModal} from '@app/helpers';
 import {Wallet} from '@app/models/wallet';
 
 export async function onAppStarted() {
-  const initialUrl = await Linking.getInitialURL();
+  const linkingSubscription = ({url}: {url: string}) => {
+    app.emit(Events.onDeepLink, url);
+  };
 
-  if (initialUrl && initialUrl.startsWith('haqq:')) {
-    app.emit(Events.onDeepLink, initialUrl);
-  }
+  Linking.addListener('url', linkingSubscription);
+
+  const subscription = ({isConnected}: NetInfoState) => {
+    isConnected ? hideModal('no-internet') : showModal('no-internet');
+  };
+
+  NetInfo.fetch().then(subscription);
+
+  NetInfo.addEventListener(subscription);
+  AppState.addEventListener('change', () => {
+    if (AppState.currentState === 'active') {
+      NetInfo.fetch().then(subscription);
+    }
+  });
+
+  const initialUrl = await Linking.getInitialURL();
+  app.emit(Events.onDeepLink, initialUrl);
 
   const wallets = Wallet.getAllVisible();
